requirements.txt
openai>=1.40.0
elevenlabs>=1.5.0
google-api-python-client>=2.140.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.8.0
